Multi page: Conditional rendering in smarty
Because all elements have to be added in PHP for the their values from the last page to be able to be read the decision for rendering them has to be moved from PHP to smarty. They only need to be added in PHP but do not need to be rendered in HTML for our purposes.